create SymbolicMathTask template
