+++ lm_eval/tasks/minerva_math.py
@@ -12,6 +12,7 @@ import re
 import math
 import code
 import signal
+from abc import ABC
 
 import sympy
 from sympy.parsing.latex import parse_latex
@@ -20,7 +21,8 @@ import inspect
 import lm_eval.datasets.hendrycks_math.hendrycks_math
 from lm_eval.metrics import mean
 from lm_eval.base import Task, rf
-from lm_eval.utils import SymbolicMathMixin, MajorityVotingMixin
+from lm_eval.utils import MajorityVotingMixin
+from lm_eval.tasks.math_tasks import SymbolicMathTask
 
 
 NL_PROMPT=r"""Problem:
@@ -83,17 +85,11 @@ _CITATION = """
 }
 """
 
-class MinervaMath(Task, SymbolicMathMixin, MajorityVotingMixin):
+
+class MinervaMath(SymbolicMathTask):
     DATASET_PATH = inspect.getfile(lm_eval.datasets.hendrycks_math.hendrycks_math)
     DATASET_NAME = None
-    MAJORITY_VOTING = "majority_voting"
-    SAMPLING_TEMPERATURE = "sampling_temperature"
-    TOP_P = "top_p"
-    EVAL_BATCH_SIZE = "eval_batch_size"
-    INVALID_ANSWER="[invalidanswer]"
     PROMPT = NL_PROMPT
-
-    end_seq = "I hope it is correct."
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
@@ -117,10 +113,30 @@ class MinervaMath(Task, SymbolicMathMixin, MajorityVotingMixin):
     def test_docs(self):
         return map(self._process_doc, self.dataset["test"])
 
-    def doc_to_target(self):
-        raise NotImplementedError("MinervaMath has no doc_to_target method.")
+    def fewshot_context(
+            self, doc, num_fewshot, provide_description=None, rnd=None, description=None
+    ):
+        example = self._doc_to_text(doc)
+        prompt = self.PROMPT + "\n\n" + example
 
-    def last_boxed_only_string(self, string):
+        return prompt
+
+    @property
+    def end_seq(self):
+        return "I hope it is correct."
+
+    def get_unnormalized_answer(self, text: str):
+        text += self.end_seq
+        match = re.search(
+                r'Final Answer: The final answer is(.*?). I hope it is correct.',
+                text,
+        )
+        if match: 
+            return match.group(1).strip()
+        else:
+            return self.INVALID_ANSWER
+
+    def _last_boxed_only_string(self, string):
 
         idx = string.rfind("\\boxed")
         if "\\boxed " in string:
@@ -150,7 +166,7 @@ class MinervaMath(Task, SymbolicMathMixin, MajorityVotingMixin):
 
         return retval
 
-    def remove_boxed(self, s):
+    def _remove_boxed(self, s):
         if "\\boxed " in s:
             left = "\\boxed "
             assert s[: len(left)] == left
@@ -165,110 +181,13 @@ class MinervaMath(Task, SymbolicMathMixin, MajorityVotingMixin):
 
     def _process_doc(self, doc):
         doc["answer"] = self.normalize_tex(
-                self.remove_boxed(self.last_boxed_only_string(doc["solution"]))
+                self._remove_boxed(self._last_boxed_only_string(doc["solution"]))
         )
         return doc
 
-    def doc_to_text(self, doc):
+    def _doc_to_text(self, doc):
         return "Problem:\n" + doc["problem"] + "\n\nSolution:"
-
-    def should_decontaminate(self):
-        return True
-
-    def doc_to_decontamination_query(self, doc):
-        return doc["problem"]
-
-    def construct_requests(self, doc, ctx, params={}):
-        if params == {}:
-            return rf.generate(ctx, [self.end_seq])
-        
-        majority_voting_value = int(params.get(self.MAJORITY_VOTING, 1))
-        sampling_temperature_value = float(params.get(self.SAMPLING_TEMPERATURE, 1.0))
-        top_p = float(params.get(self.TOP_P, 1.0))
-        eval_batch_size = params.get(self.EVAL_BATCH_SIZE, None)
-        eval_batch_size = int(eval_batch_size) if isinstance(eval_batch_size, str) else eval_batch_size
-        generation_params = {
-            'num_return_sequences': majority_voting_value,
-            'temperature': sampling_temperature_value,
-            'top_p': top_p,
-            'num_return_sequences_batch': eval_batch_size
-        }
-        return rf.generate(ctx, [self.end_seq], generation_params)
-    
-    def fewshot_context(
-            self, doc, num_fewshot, provide_description=None, rnd=None, description=None
-    ):
-        example = self.doc_to_text(doc)
-        prompt = self.PROMPT + "\n\n" + example
-
-        return prompt
-
-    def get_unnormalized_answer(self, text: str):
-        text += self.end_seq
-        match = re.search(
-                r'Final Answer: The final answer is(.*?). I hope it is correct.',
-                text,
-        )
-        if match: 
-            return match.group(1).strip()
-        else:
-            return self.INVALID_ANSWER
-
-    def process_results(self, doc, results, params={}):
-        candidates = results[0]
-
-        assert isinstance(params, dict)
-        
-        if self.MAJORITY_VOTING not in params:
-            unnormalized_answer = self.get_unnormalized_answer(candidates)
-            answer = self.normalize_tex(unnormalized_answer)
-
-            if unnormalized_answer==self.INVALID_ANSWER:
-                acc = 0
-            elif self.is_tex_equiv(answer, doc['answer']):
-                acc = 1 
-            else: 
-                acc = 0 
-
-            pass_rate = acc
-        else:
-            answers = [
-                    self.normalize_final_answer(self.get_unnormalized_answer(candidate))
-                    for candidate in candidates
-                    if self.get_unnormalized_answer(candidate) != self.INVALID_ANSWER
-            ]
-            
-            acc, pass_rate, votes = self.majority_vote(
-                    answers,
-                    correct_answer=doc['answer'],
-                    is_equiv=self.is_tex_equiv,
-            )
-            if votes:
-                answer = votes[0][0]
-            else: 
-                answer = self.INVALID_ANSWER
-
-        results = {
-            "acc": acc,
-            "pass_rate": pass_rate,
-            "metadata": {
-                "selected_answer": answer,
-                "unprocessed_answers": candidates,
-            }
-        }
-
-        if self.MAJORITY_VOTING in params:
-            results["metadata"]["votes"] = votes
-
-        return results
-
-    def aggregation(self):
-        return {"acc": mean, "pass_rate": mean, "log_pass_rate": mean}
-
-    def higher_is_better(self):
-        return {"acc": True, "pass_rate": True, "log_pass_rate": mean}
-
-
+ 
 class MinervaMathAlgebraEasy(MinervaMath):
     VERSION = 1
     DATASET_NAME = "algebra"
